Support hyper::Uri
Credit for this implementation should go to @ferranpujolcamins and @DarrenTsung.
There have been a few PRs to try and get this functionality added to hyper_serde
over the years, and the git history is scattered. I have condensed the changes
into one commit to try and make the diff easier to review/approve.